rustc_metadata: Move `CrateMetadata` into `decoder.rs`
It allows to make most of its fields private
