+++ src/librustc_metadata/cstore.rs
@@ -1,91 +1,17 @@
 // The crate store - a central repo for information collected about external
 // crates and libraries
 
-use crate::rmeta::{CrateRoot, ImportedSourceFile, Lazy, MetadataBlob};
-use rustc::dep_graph::DepNodeIndex;
-use rustc::hir::def_id::{CrateNum, DefIndex};
-use rustc::hir::map::definitions::DefPathTable;
-use rustc::middle::cstore::{CrateSource, DepKind, ExternCrate};
-use rustc::mir::interpret::AllocDecodingState;
-use rustc::session::Session;
+use crate::rmeta::CrateMetadata;
+
+use rustc_data_structures::sync::Lrc;
 use rustc_index::vec::IndexVec;
-use rustc::util::common::record_time;
-use rustc::util::nodemap::FxHashMap;
-use rustc_data_structures::sync::{Lrc, Lock, Once, AtomicCell};
-use rustc_data_structures::svh::Svh;
+use rustc::hir::def_id::CrateNum;
 use syntax::ast;
 use syntax::edition::Edition;
-use syntax_expand::base::SyntaxExtension;
 use syntax::expand::allocator::AllocatorKind;
-use proc_macro::bridge::client::ProcMacro;
+use syntax_expand::base::SyntaxExtension;
 
 pub use crate::rmeta::{provide, provide_extern};
-
-// A map from external crate numbers (as decoded from some crate file) to
-// local crate numbers (as generated during this session). Each external
-// crate may refer to types in other external crates, and each has their
-// own crate numbers.
-crate type CrateNumMap = IndexVec<CrateNum, CrateNum>;
-
-crate struct CrateMetadata {
-    /// The primary crate data - binary metadata blob.
-    crate blob: MetadataBlob,
-
-    // --- Some data pre-decoded from the metadata blob, usually for performance ---
-
-    /// Properties of the whole crate.
-    /// NOTE(eddyb) we pass `'static` to a `'tcx` parameter because this
-    /// lifetime is only used behind `Lazy`, and therefore acts like an
-    /// universal (`for<'tcx>`), that is paired up with whichever `TyCtxt`
-    /// is being used to decode those values.
-    crate root: CrateRoot<'static>,
-    /// For each definition in this crate, we encode a key. When the
-    /// crate is loaded, we read all the keys and put them in this
-    /// hashmap, which gives the reverse mapping. This allows us to
-    /// quickly retrace a `DefPath`, which is needed for incremental
-    /// compilation support.
-    crate def_path_table: DefPathTable,
-    /// Trait impl data.
-    /// FIXME: Used only from queries and can use query cache,
-    /// so pre-decoding can probably be avoided.
-    crate trait_impls: FxHashMap<(u32, DefIndex), Lazy<[DefIndex]>>,
-    /// Proc macro descriptions for this crate, if it's a proc macro crate.
-    crate raw_proc_macros: Option<&'static [ProcMacro]>,
-    /// Source maps for code from the crate.
-    crate source_map_import_info: Once<Vec<ImportedSourceFile>>,
-    /// Used for decoding interpret::AllocIds in a cached & thread-safe manner.
-    crate alloc_decoding_state: AllocDecodingState,
-    /// The `DepNodeIndex` of the `DepNode` representing this upstream crate.
-    /// It is initialized on the first access in `get_crate_dep_node_index()`.
-    /// Do not access the value directly, as it might not have been initialized yet.
-    /// The field must always be initialized to `DepNodeIndex::INVALID`.
-    crate dep_node_index: AtomicCell<DepNodeIndex>,
-
-    // --- Other significant crate properties ---
-
-    /// ID of this crate, from the current compilation session's point of view.
-    crate cnum: CrateNum,
-    /// Maps crate IDs as they are were seen from this crate's compilation sessions into
-    /// IDs as they are seen from the current compilation session.
-    crate cnum_map: CrateNumMap,
-    /// Same ID set as `cnum_map` plus maybe some injected crates like panic runtime.
-    crate dependencies: Lock<Vec<CrateNum>>,
-    /// How to link (or not link) this crate to the currently compiled crate.
-    crate dep_kind: Lock<DepKind>,
-    /// Filesystem location of this crate.
-    crate source: CrateSource,
-    /// Whether or not this crate should be consider a private dependency
-    /// for purposes of the 'exported_private_dependencies' lint
-    crate private_dep: bool,
-    /// The hash for the host proc macro. Used to support `-Z dual-proc-macro`.
-    crate host_hash: Option<Svh>,
-
-    // --- Data used only for improving diagnostics ---
-
-    /// Information about the `extern crate` item or path that caused this crate to be loaded.
-    /// If this is `None`, then the crate was injected (e.g., by the allocator).
-    crate extern_crate: Lock<Option<ExternCrate>>,
-}
 
 #[derive(Clone)]
 pub struct CStore {
@@ -97,48 +23,6 @@ pub struct CStore {
 pub enum LoadedMacro {
     MacroDef(ast::Item, Edition),
     ProcMacro(SyntaxExtension),
-}
-
-impl CrateMetadata {
-    crate fn new(
-        sess: &Session,
-        blob: MetadataBlob,
-        root: CrateRoot<'static>,
-        raw_proc_macros: Option<&'static [ProcMacro]>,
-        cnum: CrateNum,
-        cnum_map: CrateNumMap,
-        dep_kind: DepKind,
-        source: CrateSource,
-        private_dep: bool,
-        host_hash: Option<Svh>,
-    ) -> CrateMetadata {
-        let def_path_table = record_time(&sess.perf_stats.decode_def_path_tables_time, || {
-            root.def_path_table.decode((&blob, sess))
-        });
-        let trait_impls = root.impls.decode((&blob, sess))
-            .map(|trait_impls| (trait_impls.trait_id, trait_impls.impls)).collect();
-        let alloc_decoding_state =
-            AllocDecodingState::new(root.interpret_alloc_index.decode(&blob).collect());
-        let dependencies = Lock::new(cnum_map.iter().cloned().collect());
-        CrateMetadata {
-            blob,
-            root,
-            def_path_table,
-            trait_impls,
-            raw_proc_macros,
-            source_map_import_info: Once::new(),
-            alloc_decoding_state,
-            dep_node_index: AtomicCell::new(DepNodeIndex::INVALID),
-            cnum,
-            cnum_map,
-            dependencies,
-            dep_kind: Lock::new(dep_kind),
-            source,
-            private_dep,
-            host_hash,
-            extern_crate: Lock::new(None),
-        }
-    }
 }
 
 impl Default for CStore {
